+++ vql/parsers/sqlite_test.go
@@ -23,6 +23,7 @@ import (
 	vql_subsystem "www.velocidex.com/golang/velociraptor/vql"
 	vfilter "www.velocidex.com/golang/vfilter"
 
+	_ "www.velocidex.com/golang/velociraptor/accessors/file"
 	_ "www.velocidex.com/golang/velociraptor/result_sets/timed"
 )
