+++ vql/readers/paged_reader_test.go
@@ -17,6 +17,8 @@ import (
 	vql_subsystem "www.velocidex.com/golang/velociraptor/vql"
 	"www.velocidex.com/golang/velociraptor/vtesting"
 	"www.velocidex.com/golang/vfilter"
+
+	_ "www.velocidex.com/golang/velociraptor/accessors/file"
 )
 
 type TestSuite struct {
@@ -31,6 +33,7 @@ func (self *TestSuite) SetupTest() {
 	self.scope = vql_subsystem.MakeScope()
 	self.scope.AppendVars(ordereddict.NewDict().
 		Set(vql_subsystem.CACHE_VAR, vql_subsystem.NewScopeCache()).
+		Set(vql_subsystem.ACL_MANAGER_VAR, vql_subsystem.NullACLManager{}).
 		Set(constants.SCOPE_ROOT, self.scope))
 
 	// Make a very small pool
@@ -72,7 +75,8 @@ func (self *TestSuite) TestPagedReader() {
 	for i := 0; i < 10; i++ {
 		reader, err := NewPagedReader(self.scope, "file", self.filenames[i], 100)
 		assert.NoError(self.T(), err)
-		reader.ReadAt(buff, 0)
+		_, err = reader.ReadAt(buff, 0)
+		assert.NoError(self.T(), err)
 		assert.Equal(self.T(), binary.LittleEndian.Uint32(buff), uint32(i))
 		readers = append(readers, reader)
 	}
@@ -80,7 +84,8 @@ func (self *TestSuite) TestPagedReader() {
 	for i := 0; i < 10; i++ {
 		reader, err := NewPagedReader(self.scope, "file", self.filenames[i], 100)
 		assert.NoError(self.T(), err)
-		reader.ReadAt(buff, 0)
+		_, err = reader.ReadAt(buff, 0)
+		assert.NoError(self.T(), err)
 		assert.Equal(self.T(), binary.LittleEndian.Uint32(buff), uint32(i))
 	}
 
@@ -89,7 +94,9 @@ func (self *TestSuite) TestPagedReader() {
 		reader, err := NewPagedReader(self.scope, "file", self.filenames[1], 100)
 		assert.NoError(self.T(), err)
 
-		reader.ReadAt(buff, 0)
+		_, err = reader.ReadAt(buff, 0)
+		assert.NoError(self.T(), err)
+
 		assert.Equal(self.T(), binary.LittleEndian.Uint32(buff), uint32(1))
 	}
 
@@ -100,14 +107,17 @@ func (self *TestSuite) TestPagedReader() {
 
 	for i := 0; i < 10; i++ {
 		reader.Close()
-		reader.ReadAt(buff, 0)
+		_, err = reader.ReadAt(buff, 0)
+		assert.NoError(self.T(), err)
+
 		assert.Equal(self.T(), binary.LittleEndian.Uint32(buff), uint32(1))
 	}
 
 	// Make the reader's timeout very short.
 	reader.SetLifetime(10 * time.Millisecond)
 	reader.Close()
-	reader.ReadAt(buff, 0)
+	_, err = reader.ReadAt(buff, 0)
+	assert.NoError(self.T(), err)
 	assert.Equal(self.T(), binary.LittleEndian.Uint32(buff), uint32(1))
 
 	// Wait here until the reader closes itself by itself.
@@ -119,7 +129,8 @@ func (self *TestSuite) TestPagedReader() {
 	})
 
 	// Next read still works.
-	reader.ReadAt(buff, 0)
+	_, err = reader.ReadAt(buff, 0)
+	assert.NoError(self.T(), err)
 	assert.Equal(self.T(), binary.LittleEndian.Uint32(buff), uint32(1))
 
 	// Close the scope - this should close all the pool
